+++ deploy/crds/multicloud-apps.io_placementrules_crd.yaml
@@ -1,13 +1,13 @@
 apiVersion: apiextensions.k8s.io/v1beta1
 kind: CustomResourceDefinition
 metadata:
-  name: placementrules.multicloud.io
+  name: placementrules.multicloud-apps.io
 spec:
   additionalPrinterColumns:
   - JSONPath: .metadata.creationTimestamp
     name: Age
     type: date
-  group: multicloud.io
+  group: multicloud-apps.io
   names:
     kind: PlacementRule
     listKind: PlacementRuleList
