+++ examples/all-clusters-placement.yaml
@@ -1,4 +1,4 @@
-apiVersion: multicloud.io/v1alpha1
+apiVersion: multicloud-apps.io/v1alpha1
 kind: PlacementRule
 metadata:
   name: allclusters
