+++ pkg/apis/multicloudapps/v1alpha1/register.go
@@ -14,9 +14,9 @@
 
 // NOTE: Boilerplate only.  Ignore this file.
 
-// Package v1alpha1 contains API Schema definitions for the multicloud v1alpha1 API group
+// Package v1alpha1 contains API Schema definitions for the multicloud-apps v1alpha1 API group
 // +k8s:deepcopy-gen=package,register
-// +groupName=multicloud.io
+// +groupName=multicloud-apps.io
 package v1alpha1
 
 import (
@@ -26,7 +26,7 @@ import (
 
 var (
 	// SchemeGroupVersion is group version used to register these objects
-	SchemeGroupVersion = schema.GroupVersion{Group: "multicloud.io", Version: "v1alpha1"}
+	SchemeGroupVersion = schema.GroupVersion{Group: "multicloud-apps.io", Version: "v1alpha1"}
 
 	// SchemeBuilder is used to add go types to the GroupVersionKind scheme
 	SchemeBuilder = &scheme.Builder{GroupVersion: SchemeGroupVersion}
