+++ pkg/utils/utils_test.go
@@ -54,7 +54,7 @@ func TestLoadCRD(t *testing.T) {
 	g := gomega.NewGomegaWithT(t)
 
 	g.Expect(CheckAndInstallCRD(cfg, "../../deploy/crds/app.ibm.com_placementrules_crd.yaml")).NotTo(gomega.HaveOccurred())
-	g.Expect(CheckAndInstallCRD(cfg, "../../deploy/crds/multicloud.io_placementrules_crd.yaml")).NotTo(gomega.HaveOccurred())
+	g.Expect(CheckAndInstallCRD(cfg, "../../deploy/crds/multicloud-apps.io_placementrules_crd.yaml")).NotTo(gomega.HaveOccurred())
 }
 
 func TestEventRecorder(t *testing.T) {
